+++ changelog.md
@@ -23,7 +23,7 @@
 ## 4.0.0 - 2017-05-16
 
 * Allow for google/cloud ^0.26.0|^0.27.0|^0.28.0
-* Bump up to superbalist/php-pubsub ^2.0
+* Bump up to milind/php-pubsub ^2.0
 * Add new publishBatch method to GoogleCloudPubSubAdapter
 
 ## 3.0.1 - 2017-04-03